move test to correct package
